refactor(async-await): migration on MockTokenizationService
