+++ PrimerSDK.xcworkspace/contents.xcworkspacedata
@@ -2559,7 +2559,7 @@
                   location = "group:MockPrimerUIManager.swift">
                </FileRef>
                <FileRef
-                  location = "group:TokenizationService.swift">
+                  location = "group:MockTokenizationService.swift">
                </FileRef>
             </Group>
             <Group
